Rendre riktig ikon basert på hendelse
